add: binarySearch fourth iteration
added logging package to log application behaviour. It is understood that testing logs and log package (from stdlib) direct their outputs to different 'channels', for instance, tests logs output are directed from the test runner. Specified the main package to get logs from application itself with a 'simplified' application of the topic in question, in this case the binary search algorithm. In order to get logs from the testcase one may need to apply verbose mode when calling tests, but depending on the test suite, this may overwhelm with a lot more messages per test (even the ones that pass).